Gateway API specifics decoupled from Topology
