+++ machinery/gateway_api_types.go
@@ -1,22 +1,21 @@
 package machinery
 
 import (
+	"fmt"
+
 	core "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/utils/ptr"
 	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
+	gwapiv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 )
 
-// Targetable is an interface that represents an object that can be targeted by policies.
-type Targetable interface {
-	Object
+const nameSectionNameURLSeparator = '#'
 
-	SetPolicies([]Policy)
-	Policies() []Policy
-}
-
-func MapTargetableToURLFunc(t Targetable, _ int) string {
-	return t.GetURL()
-}
+// These are wrappers for Gateway API types so instances can be used as targetables in the topology.
+// Targateables typically store back references to the policies that are attached to them.
+// The implementation of GetURL() must return a unique identifier for the wrapped object that matches the one
+// generated by policy targetRefs that implement the PolicyTargetReference interface for values pointing to the object.
 
 type GatewayClass struct {
 	*gwapiv1.GatewayClass
@@ -212,4 +211,110 @@ func (p *ServicePort) SetPolicies(policies []Policy) {
 
 func (p *ServicePort) Policies() []Policy {
 	return p.attachedPolicies
+}
+
+// These are Gateway API target reference types that implement the PolicyTargetReference interface, so policies'
+// targetRef instances can be treated as Objects whose GetURL() functions return the unique identifier of the
+// corresponding targetable the reference points to.
+// This is the reason why GetURL() was adopted to get the unique identifiers of topology objects instead of more
+// obvious Kubernetes objects' GetUID() (k8s.io/apimachinery/pkg/apis/meta/v1).
+
+type NamespacedPolicyTargetReference struct {
+	gwapiv1alpha2.NamespacedPolicyTargetReference
+	PolicyNamespace string
+}
+
+var _ PolicyTargetReference = NamespacedPolicyTargetReference{}
+
+func (t NamespacedPolicyTargetReference) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group: string(t.Group),
+		Kind:  string(t.Kind),
+	}
+}
+
+func (t NamespacedPolicyTargetReference) SetGroupVersionKind(gvk schema.GroupVersionKind) {
+	t.Group = gwapiv1alpha2.Group(gvk.Group)
+	t.Kind = gwapiv1alpha2.Kind(gvk.Kind)
+}
+
+func (t NamespacedPolicyTargetReference) GetURL() string {
+	return UrlFromObject(t)
+}
+
+func (t NamespacedPolicyTargetReference) GetNamespace() string {
+	return string(ptr.Deref(t.Namespace, gwapiv1alpha2.Namespace(t.PolicyNamespace)))
+}
+
+func (t NamespacedPolicyTargetReference) GetName() string {
+	return string(t.NamespacedPolicyTargetReference.Name)
+}
+
+type LocalPolicyTargetReference struct {
+	gwapiv1alpha2.LocalPolicyTargetReference
+	PolicyNamespace string
+}
+
+var _ PolicyTargetReference = LocalPolicyTargetReference{}
+
+func (t LocalPolicyTargetReference) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group: string(t.Group),
+		Kind:  string(t.Kind),
+	}
+}
+
+func (t LocalPolicyTargetReference) SetGroupVersionKind(gvk schema.GroupVersionKind) {
+	t.Group = gwapiv1alpha2.Group(gvk.Group)
+	t.Kind = gwapiv1alpha2.Kind(gvk.Kind)
+}
+
+func (t LocalPolicyTargetReference) GetURL() string {
+	return UrlFromObject(t)
+}
+
+func (t LocalPolicyTargetReference) GetNamespace() string {
+	return t.PolicyNamespace
+}
+
+func (t LocalPolicyTargetReference) GetName() string {
+	return string(t.LocalPolicyTargetReference.Name)
+}
+
+type LocalPolicyTargetReferenceWithSectionName struct {
+	gwapiv1alpha2.LocalPolicyTargetReferenceWithSectionName
+	PolicyNamespace string
+}
+
+var _ PolicyTargetReference = LocalPolicyTargetReferenceWithSectionName{}
+
+func (t LocalPolicyTargetReferenceWithSectionName) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group: string(t.Group),
+		Kind:  string(t.Kind),
+	}
+}
+
+func (t LocalPolicyTargetReferenceWithSectionName) SetGroupVersionKind(gvk schema.GroupVersionKind) {
+	t.Group = gwapiv1alpha2.Group(gvk.Group)
+	t.Kind = gwapiv1alpha2.Kind(gvk.Kind)
+}
+
+func (t LocalPolicyTargetReferenceWithSectionName) GetURL() string {
+	return UrlFromObject(t)
+}
+
+func (t LocalPolicyTargetReferenceWithSectionName) GetNamespace() string {
+	return t.PolicyNamespace
+}
+
+func (t LocalPolicyTargetReferenceWithSectionName) GetName() string {
+	if t.SectionName == nil {
+		return string(t.LocalPolicyTargetReference.Name)
+	}
+	return namespacedSectionName(string(t.LocalPolicyTargetReference.Name), *t.SectionName)
+}
+
+func namespacedSectionName(namespace string, sectionName gwapiv1.SectionName) string {
+	return fmt.Sprintf("%s%s%s", namespace, string(nameSectionNameURLSeparator), sectionName)
 }
