+++ machinery/test_helper.go
@@ -3,374 +3,280 @@
 package machinery
 
 import (
-	core "k8s.io/api/core/v1"
+	"bytes"
+	"fmt"
+	"os"
+	"strings"
+	"testing"
+
+	"github.com/samber/lo"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/utils/ptr"
-	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
 )
 
-func BuildGatewayClass(f ...func(*gwapiv1.GatewayClass)) *gwapiv1.GatewayClass {
-	gc := &gwapiv1.GatewayClass{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: gwapiv1.GroupVersion.String(),
-			Kind:       "GatewayClass",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "my-gateway-class",
-		},
-		Spec: gwapiv1.GatewayClassSpec{
-			ControllerName: gwapiv1.GatewayController("my-gateway-controller"),
-		},
+// SaveToOutputDir saves the output of a test case to a file in the output directory.
+func SaveToOutputDir(t *testing.T, out *bytes.Buffer, outDir, ext string) {
+	file, err := os.Create(fmt.Sprintf("%s/%s%s", outDir, strings.ReplaceAll(t.Name(), "/", "__"), ext))
+	if err != nil {
+		t.Fatal(err)
 	}
-	for _, fn := range f {
-		fn(gc)
+	defer file.Close()
+	_, err = file.Write(out.Bytes())
+	if err != nil {
+		t.Fatal(err)
 	}
-	return gc
 }
 
-func BuildGateway(f ...func(*gwapiv1.Gateway)) *gwapiv1.Gateway {
-	g := &gwapiv1.Gateway{
+func linksFromNode(topology *Topology, node Targetable, edges map[string][]string) {
+	if _, ok := edges[node.GetName()]; ok {
+		return
+	}
+	children := topology.Children(node)
+	edges[node.GetName()] = lo.Map(children, func(child Targetable, _ int) string { return child.GetName() })
+	for _, child := range children {
+		linksFromNode(topology, child, edges)
+	}
+}
+
+const TestGroupName = "example.test"
+
+type Apple struct {
+	Name string
+
+	policies []Policy
+}
+
+var _ Targetable = &Apple{}
+
+func (a *Apple) GetName() string {
+	return a.Name
+}
+
+func (a *Apple) GetNamespace() string {
+	return ""
+}
+
+func (a *Apple) GetURL() string {
+	return UrlFromObject(a)
+}
+
+func (a *Apple) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group:   TestGroupName,
+		Version: "v1",
+		Kind:    "Apple",
+	}
+}
+
+func (a *Apple) SetGroupVersionKind(schema.GroupVersionKind) {}
+
+func (a *Apple) Policies() []Policy {
+	return a.policies
+}
+
+func (a *Apple) SetPolicies(policies []Policy) {
+	a.policies = policies
+}
+
+type Orange struct {
+	Name         string
+	Namespace    string
+	AppleParents []string
+	ChildBananas []string
+
+	policies []Policy
+}
+
+var _ Targetable = &Orange{}
+
+func (o *Orange) GetName() string {
+	return o.Name
+}
+
+func (o *Orange) GetNamespace() string {
+	return o.Namespace
+}
+
+func (o *Orange) GetURL() string {
+	return UrlFromObject(o)
+}
+
+func (o *Orange) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group:   TestGroupName,
+		Version: "v1beta1",
+		Kind:    "Orange",
+	}
+}
+
+func (o *Orange) SetGroupVersionKind(schema.GroupVersionKind) {}
+
+func (o *Orange) Policies() []Policy {
+	return o.policies
+}
+
+func (o *Orange) SetPolicies(policies []Policy) {
+	o.policies = policies
+}
+
+type Banana struct {
+	Name string
+}
+
+var _ Targetable = &Banana{}
+
+func (b *Banana) GetName() string {
+	return b.Name
+}
+
+func (b *Banana) GetNamespace() string {
+	return ""
+}
+
+func (b *Banana) GetURL() string {
+	return UrlFromObject(b)
+}
+
+func (b *Banana) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group:   TestGroupName,
+		Version: "v1beta1",
+		Kind:    "Banana",
+	}
+}
+
+func (b *Banana) SetGroupVersionKind(schema.GroupVersionKind) {}
+
+func (b *Banana) Policies() []Policy {
+	return nil
+}
+
+func (b *Banana) SetPolicies(policies []Policy) {}
+
+func LinkApplesToOranges(apples []*Apple) LinkFunc {
+	return LinkFunc{
+		From: schema.GroupKind{Group: TestGroupName, Kind: "Apple"},
+		To:   schema.GroupKind{Group: TestGroupName, Kind: "Orange"},
+		Func: func(child Targetable) []Targetable {
+			orange := child.(*Orange)
+			return lo.FilterMap(apples, func(apple *Apple, _ int) (Targetable, bool) {
+				return apple, lo.Contains(orange.AppleParents, apple.Name)
+			})
+		},
+	}
+}
+
+func LinkOrangesToBananas(oranges []*Orange) LinkFunc {
+	return LinkFunc{
+		From: schema.GroupKind{Group: TestGroupName, Kind: "Orange"},
+		To:   schema.GroupKind{Group: TestGroupName, Kind: "Banana"},
+		Func: func(child Targetable) []Targetable {
+			banana := child.(*Banana)
+			return lo.FilterMap(oranges, func(orange *Orange, _ int) (Targetable, bool) {
+				return orange, lo.Contains(orange.ChildBananas, banana.Name)
+			})
+		},
+	}
+}
+
+type FruitPolicy struct {
+	metav1.TypeMeta   `json:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+
+	Spec FruitPolicySpec `json:"spec"`
+}
+
+type FruitPolicySpec struct {
+	TargetRef FruitPolicyTargetReference `json:"targetRef"`
+}
+
+var _ Policy = &FruitPolicy{}
+
+func (p *FruitPolicy) GetURL() string {
+	return UrlFromObject(p)
+}
+
+func (p *FruitPolicy) GetTargetRefs() []PolicyTargetReference {
+	var namespace *string
+	group := p.Spec.TargetRef.Group
+	kind := p.Spec.TargetRef.Kind
+	if group == TestGroupName && kind == "Orange" {
+		namespace = ptr.To(ptr.Deref(p.Spec.TargetRef.Namespace, p.Namespace))
+	}
+	return []PolicyTargetReference{
+		FruitPolicyTargetReference{
+			Group:     group,
+			Kind:      kind,
+			Name:      p.Spec.TargetRef.Name,
+			Namespace: namespace,
+		},
+	}
+}
+
+func (p *FruitPolicy) GetMergeStrategy() MergeStrategy {
+	return DefaultMergeStrategy
+}
+
+func (p *FruitPolicy) Merge(policy Policy) Policy {
+	return &FruitPolicy{
+		Spec: p.Spec,
+	}
+}
+
+type FruitPolicyTargetReference struct {
+	Group     string
+	Kind      string
+	Name      string
+	Namespace *string
+}
+
+var _ PolicyTargetReference = FruitPolicyTargetReference{}
+
+func (t FruitPolicyTargetReference) GroupVersionKind() schema.GroupVersionKind {
+	return schema.GroupVersionKind{
+		Group: t.Group,
+		Kind:  t.Kind,
+	}
+}
+
+func (t FruitPolicyTargetReference) SetGroupVersionKind(gvk schema.GroupVersionKind) {
+	t.Group = gvk.Group
+	t.Kind = gvk.Kind
+}
+
+func (t FruitPolicyTargetReference) GetURL() string {
+	return UrlFromObject(t)
+}
+
+func (t FruitPolicyTargetReference) GetNamespace() string {
+	return ptr.Deref(t.Namespace, "")
+}
+
+func (t FruitPolicyTargetReference) GetName() string {
+	return t.Name
+}
+
+func buildFruitPolicy(f ...func(*FruitPolicy)) *FruitPolicy {
+	p := &FruitPolicy{
 		TypeMeta: metav1.TypeMeta{
-			APIVersion: gwapiv1.GroupVersion.String(),
-			Kind:       "Gateway",
+			APIVersion: "test/v1",
+			Kind:       "FruitPolicy",
 		},
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      "my-gateway",
+			Name:      "my-policy",
 			Namespace: "my-namespace",
 		},
-		Spec: gwapiv1.GatewaySpec{
-			GatewayClassName: "my-gateway-class",
-			Listeners: []gwapiv1.Listener{
-				{
-					Name:     "my-listener",
-					Port:     80,
-					Protocol: "HTTP",
-				},
+		Spec: FruitPolicySpec{
+			TargetRef: FruitPolicyTargetReference{
+				Group: TestGroupName,
+				Kind:  "Orange",
+				Name:  "my-orange",
 			},
 		},
 	}
 	for _, fn := range f {
-		fn(g)
+		fn(p)
 	}
-	return g
-}
-
-func BuildHTTPRoute(f ...func(*gwapiv1.HTTPRoute)) *gwapiv1.HTTPRoute {
-	r := &gwapiv1.HTTPRoute{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: gwapiv1.GroupVersion.String(),
-			Kind:       "HTTPRoute",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "my-http-route",
-			Namespace: "my-namespace",
-		},
-		Spec: gwapiv1.HTTPRouteSpec{
-			CommonRouteSpec: gwapiv1.CommonRouteSpec{
-				ParentRefs: []gwapiv1.ParentReference{
-					{
-						Name: "my-gateway",
-					},
-				},
-			},
-			Rules: []gwapiv1.HTTPRouteRule{
-				{
-					BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef()},
-				},
-			},
-		},
-	}
-	for _, fn := range f {
-		fn(r)
-	}
-	return r
-}
-
-func BuildHTTPBackendRef(f ...func(*gwapiv1.BackendObjectReference)) gwapiv1.HTTPBackendRef {
-	bor := &gwapiv1.BackendObjectReference{
-		Name: "my-service",
-	}
-	for _, fn := range f {
-		fn(bor)
-	}
-	return gwapiv1.HTTPBackendRef{
-		BackendRef: gwapiv1.BackendRef{
-			BackendObjectReference: *bor,
-		},
-	}
-}
-
-func BuildService(f ...func(*core.Service)) *core.Service {
-	s := &core.Service{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: core.SchemeGroupVersion.String(),
-			Kind:       "Service",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "my-service",
-			Namespace: "my-namespace",
-		},
-		Spec: core.ServiceSpec{
-			Ports: []core.ServicePort{
-				{
-					Name: "http",
-					Port: 80,
-				},
-			},
-			Selector: map[string]string{
-				"app": "my-app",
-			},
-		},
-	}
-	for _, fn := range f {
-		fn(s)
-	}
-	return s
-}
-
-type GatewayAPIResources struct {
-	GatewayClasses []*gwapiv1.GatewayClass
-	Gateways       []*gwapiv1.Gateway
-	HTTPRoutes     []*gwapiv1.HTTPRoute
-	Services       []*core.Service
-}
-
-// BuildComplexGatewayAPITopology returns a set of Gateway API resources organized :
-//
-//	                                            ┌────────────────┐                                                                        ┌────────────────┐
-//	                                            │ gatewayclass-1 │                                                                        │ gatewayclass-2 │
-//	                                            └────────────────┘                                                                        └────────────────┘
-//	                                                    ▲                                                                                         ▲
-//	                                                    │                                                                                         │
-//	                          ┌─────────────────────────┼──────────────────────────┐                                                 ┌────────────┴─────────────┐
-//	                          │                         │                          │                                                 │                          │
-//	          ┌───────────────┴───────────────┐ ┌───────┴────────┐ ┌───────────────┴───────────────┐                  ┌──────────────┴────────────────┐ ┌───────┴────────┐
-//	          │           gateway-1           │ │   gateway-2    │ │           gateway-3           │                  │           gateway-4           │ │   gateway-5    │
-//	          │                               │ │                │ │                               │                  │                               │ │                │
-//	          │ ┌────────────┐ ┌────────────┐ │ │ ┌────────────┐ │ │ ┌────────────┐ ┌────────────┐ │                  │ ┌────────────┐ ┌────────────┐ │ │ ┌────────────┐ │
-//	          │ │ listener-1 │ │ listener-2 │ │ │ │ listener-1 │ │ │ │ listener-1 │ │ listener-2 │ │                  │ │ listener-1 │ │ listener-2 │ │ │ │ listener-1 │ │
-//	          │ └────────────┘ └────────────┘ │ │ └────────────┘ │ │ └────────────┘ └────────────┘ │                  │ └────────────┘ └────────────┘ │ │ └────────────┘ │
-//	          │                        ▲      │ │      ▲         │ │                               │                  │                               │ │                │
-//	          └────────────────────────┬──────┘ └──────┬─────────┘ └───────────────────────────────┘                  └───────────────────────────────┘ └────────────────┘
-//	                      ▲            │               │     ▲                    ▲            ▲                          ▲           ▲                          ▲
-//	                      │            │               │     │                    │            │                          │           │                          │
-//	                      │            └───────┬───────┘     │                    │            └────────────┬─────────────┘           │                          │
-//	                      │                    │             │                    │                         │                         │                          │
-//	          ┌───────────┴───────────┐ ┌──────┴─────┐ ┌─────┴──────┐ ┌───────────┴───────────┐ ┌───────────┴───────────┐ ┌───────────┴───────────┐        ┌─────┴──────┐
-//	          │        route-1        │ │  route-2   │ │  route-3   │ │        route-4        │ │        route-5        │ │        route-6        │        │   route-7  │
-//	          │                       │ │            │ │            │ │                       │ │                       │ │                       │        │            │
-//	          │ ┌────────┐ ┌────────┐ │ │ ┌────────┐ │ │ ┌────────┐ │ │ ┌────────┐ ┌────────┐ │ │ ┌────────┐ ┌────────┐ │ │ ┌────────┐ ┌────────┐ │        │ ┌────────┐ │
-//	          │ │ rule-1 │ │ rule-2 │ │ │ │ rule-1 │ │ │ │ rule-1 │ │ │ │ rule-1 │ │ rule-2 │ │ │ │ rule-1 │ │ rule-2 │ │ │ │ rule-1 │ │ rule-2 │ │        │ │ rule-1 │ │
-//	          │ └────┬───┘ └────┬───┘ │ │ └────┬───┘ │ │ └───┬────┘ │ │ └─┬──────┘ └───┬────┘ │ │ └───┬────┘ └────┬───┘ │ │ └─┬────┬─┘ └────┬───┘ │        │ └────┬───┘ │
-//	          │      │          │     │ │      │     │ │     │      │ │   │            │      │ │     │           │     │ │   │    │        │     │        │      │     │
-//	          └──────┼──────────┼─────┘ └──────┼─────┘ └─────┼──────┘ └───┼────────────┼──────┘ └─────┼───────────┼─────┘ └───┼────┼────────┼─────┘        └──────┼─────┘
-//	                 │          │              │             │            │            │              │           │           │    │        │                     │
-//	                 │          │              └─────────────┤            │            │              └───────────┴───────────┘    │        │                     │
-//	                 ▼          ▼                            │            │            │                          ▼                ▼        │                     ▼
-//	┌───────────────────────┐ ┌────────────┐          ┌──────┴────────────┴───┐  ┌─────┴──────┐             ┌────────────┐        ┌─────────┴──┐           ┌────────────┐
-//	│                       │ │            │          │      ▼            ▼   │  │     ▼      │             │            │        │         ▼  │           │            │
-//	│ ┌────────┐ ┌────────┐ │ │ ┌────────┐ │          │ ┌────────┐ ┌────────┐ │  │ ┌────────┐ │             │ ┌────────┐ │        │ ┌────────┐ │           │ ┌────────┐ │
-//	│ │ port-1 │ │ port-2 │ │ │ │ port-1 │ │          │ │ port-1 │ │ port-2 │ │  │ │ port-1 │ │             │ │ port-1 │ │        │ │ port-1 │ │           │ │ port-1 │ │
-//	│ └────────┘ └────────┘ │ │ └────────┘ │          │ └────────┘ └────────┘ │  │ └────────┘ │             │ └────────┘ │        │ └────────┘ │           │ └────────┘ │
-//	│                       │ │            │          │                       │  │            │             │            │        │            │           │            │
-//	│       service-1       │ │  service-2 │          │       service-3       │  │  service-4 │             │  service-5 │        │  service-6 │           │  service-7 │
-//	└───────────────────────┘ └────────────┘          └───────────────────────┘  └────────────┘             └────────────┘        └────────────┘           └────────────┘
-func BuildComplexGatewayAPITopology(funcs ...func(*GatewayAPIResources)) GatewayAPIResources {
-	t := GatewayAPIResources{
-		GatewayClasses: []*gwapiv1.GatewayClass{
-			BuildGatewayClass(func(gc *gwapiv1.GatewayClass) { gc.Name = "gatewayclass-1" }),
-			BuildGatewayClass(func(gc *gwapiv1.GatewayClass) { gc.Name = "gatewayclass-2" }),
-		},
-		Gateways: []*gwapiv1.Gateway{
-			BuildGateway(func(g *gwapiv1.Gateway) {
-				g.Name = "gateway-1"
-				g.Spec.GatewayClassName = "gatewayclass-1"
-				g.Spec.Listeners[0].Name = "listener-1"
-				g.Spec.Listeners = append(g.Spec.Listeners, gwapiv1.Listener{
-					Name:     "listener-2",
-					Port:     443,
-					Protocol: "HTTPS",
-				})
-			}),
-			BuildGateway(func(g *gwapiv1.Gateway) {
-				g.Name = "gateway-2"
-				g.Spec.GatewayClassName = "gatewayclass-1"
-				g.Spec.Listeners[0].Name = "listener-1"
-			}),
-			BuildGateway(func(g *gwapiv1.Gateway) {
-				g.Name = "gateway-3"
-				g.Spec.GatewayClassName = "gatewayclass-1"
-				g.Spec.Listeners[0].Name = "listener-1"
-				g.Spec.Listeners = append(g.Spec.Listeners, gwapiv1.Listener{
-					Name:     "listener-2",
-					Port:     443,
-					Protocol: "HTTPS",
-				})
-			}),
-			BuildGateway(func(g *gwapiv1.Gateway) {
-				g.Name = "gateway-4"
-				g.Spec.GatewayClassName = "gatewayclass-2"
-				g.Spec.Listeners[0].Name = "listener-1"
-				g.Spec.Listeners = append(g.Spec.Listeners, gwapiv1.Listener{
-					Name:     "listener-2",
-					Port:     443,
-					Protocol: "HTTPS",
-				})
-			}),
-			BuildGateway(func(g *gwapiv1.Gateway) {
-				g.Name = "gateway-5"
-				g.Spec.GatewayClassName = "gatewayclass-2"
-				g.Spec.Listeners[0].Name = "listener-1"
-			}),
-		},
-		HTTPRoutes: []*gwapiv1.HTTPRoute{
-			BuildHTTPRoute(func(r *gwapiv1.HTTPRoute) {
-				r.Name = "route-1"
-				r.Spec.ParentRefs[0].Name = "gateway-1"
-				r.Spec.Rules = []gwapiv1.HTTPRouteRule{
-					{ // rule-1
-						BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-							backendRef.Name = "service-1"
-						})},
-					},
-					{ // rule-2
-						BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-							backendRef.Name = "service-2"
-						})},
-					},
-				}
-			}),
-			BuildHTTPRoute(func(r *gwapiv1.HTTPRoute) {
-				r.Name = "route-2"
-				r.Spec.ParentRefs = []gwapiv1.ParentReference{
-					{
-						Name:        "gateway-1",
-						SectionName: ptr.To(gwapiv1.SectionName("listener-2")),
-					},
-					{
-						Name:        "gateway-2",
-						SectionName: ptr.To(gwapiv1.SectionName("listener-1")),
-					},
-				}
-				r.Spec.Rules[0].BackendRefs[0] = BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-					backendRef.Name = "service-3"
-					backendRef.Port = ptr.To(gwapiv1.PortNumber(80)) // port-1
-				})
-			}),
-			BuildHTTPRoute(func(r *gwapiv1.HTTPRoute) {
-				r.Name = "route-3"
-				r.Spec.ParentRefs[0].Name = "gateway-2"
-				r.Spec.Rules[0].BackendRefs[0] = BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-					backendRef.Name = "service-3"
-					backendRef.Port = ptr.To(gwapiv1.PortNumber(80)) // port-1
-				})
-			}),
-			BuildHTTPRoute(func(r *gwapiv1.HTTPRoute) {
-				r.Name = "route-4"
-				r.Spec.ParentRefs[0].Name = "gateway-3"
-				r.Spec.Rules = []gwapiv1.HTTPRouteRule{
-					{ // rule-1
-						BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-							backendRef.Name = "service-3"
-							backendRef.Port = ptr.To(gwapiv1.PortNumber(443)) // port-2
-						})},
-					},
-					{ // rule-2
-						BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-							backendRef.Name = "service-4"
-							backendRef.Port = ptr.To(gwapiv1.PortNumber(80)) // port-1
-						})},
-					},
-				}
-			}),
-			BuildHTTPRoute(func(r *gwapiv1.HTTPRoute) {
-				r.Name = "route-5"
-				r.Spec.ParentRefs[0].Name = "gateway-3"
-				r.Spec.ParentRefs = append(r.Spec.ParentRefs, gwapiv1.ParentReference{Name: "gateway-4"})
-				r.Spec.Rules = []gwapiv1.HTTPRouteRule{
-					{ // rule-1
-						BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-							backendRef.Name = "service-5"
-						})},
-					},
-					{ // rule-2
-						BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-							backendRef.Name = "service-5"
-						})},
-					},
-				}
-			}),
-			BuildHTTPRoute(func(r *gwapiv1.HTTPRoute) {
-				r.Name = "route-6"
-				r.Spec.ParentRefs[0].Name = "gateway-4"
-				r.Spec.Rules = []gwapiv1.HTTPRouteRule{
-					{ // rule-1
-						BackendRefs: []gwapiv1.HTTPBackendRef{
-							BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-								backendRef.Name = "service-5"
-							}),
-							BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-								backendRef.Name = "service-6"
-							}),
-						},
-					},
-					{ // rule-2
-						BackendRefs: []gwapiv1.HTTPBackendRef{BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-							backendRef.Name = "service-6"
-							backendRef.Port = ptr.To(gwapiv1.PortNumber(80)) // port-1
-						})},
-					},
-				}
-			}),
-			BuildHTTPRoute(func(r *gwapiv1.HTTPRoute) {
-				r.Name = "route-7"
-				r.Spec.ParentRefs[0].Name = "gateway-5"
-				r.Spec.Rules[0].BackendRefs[0] = BuildHTTPBackendRef(func(backendRef *gwapiv1.BackendObjectReference) {
-					backendRef.Name = "service-7"
-				})
-			}),
-		},
-		Services: []*core.Service{
-			BuildService(func(s *core.Service) {
-				s.Name = "service-1"
-				s.Spec.Ports[0].Name = "port-1"
-				s.Spec.Ports = append(s.Spec.Ports, core.ServicePort{
-					Name: "port-2",
-					Port: 443,
-				})
-			}),
-			BuildService(func(s *core.Service) {
-				s.Name = "service-2"
-				s.Spec.Ports[0].Name = "port-1"
-			}),
-			BuildService(func(s *core.Service) {
-				s.Name = "service-3"
-				s.Spec.Ports[0].Name = "port-1"
-				s.Spec.Ports = append(s.Spec.Ports, core.ServicePort{
-					Name: "port-2",
-					Port: 443,
-				})
-			}),
-			BuildService(func(s *core.Service) {
-				s.Name = "service-4"
-				s.Spec.Ports[0].Name = "port-1"
-			}),
-			BuildService(func(s *core.Service) {
-				s.Name = "service-5"
-				s.Spec.Ports[0].Name = "port-1"
-			}),
-			BuildService(func(s *core.Service) {
-				s.Name = "service-6"
-				s.Spec.Ports[0].Name = "port-1"
-			}),
-			BuildService(func(s *core.Service) {
-				s.Name = "service-7"
-				s.Spec.Ports[0].Name = "port-1"
-			}),
-		},
-	}
-	for _, f := range funcs {
-		f(&t)
-	}
-	return t
+	return p
 }
